add support for daisy rev5
Add dynamic support for rev5 daisy board. An alternative would have been
a feature flag, but since I have both Rev4 and Rev5 boards, the runtime
detection is more convenient.

The version is detected using the hardware version pin (PD3). On Rev4,
we still use the AK5456 codec:

  SAI_*_A : master
  SAI_SD_A : STM32->DAC
  SAI_SD_B : DAC->STM32

On Rev5, another codec is used (WM8731), and a first difference is that
pins are reversed:

  SAI_*_A : master
  SAI_SD_A : DAC->STM32
  SAI_SD_B : STM32->DAC

We create 2 enums (DmaInputStreamGeneric and DmaOutputStreamGeneric) to
handle the 2 cases.

We also need to initialize the codec through I2C. The registers are
described in the datasheet (see link). The codec initialization is moved
from GPIO module to the audio module, and a function is added for each
codec. This creates a dependency: GPIOs must be initialized before audio.

Note: this commit is partially inspired from Max Genson's work,
changelog here (see 2nd link for commits):

- WM8731 codec now is configurable via I2C
- Codec outputs sound with unwanted distortion
- no distortion in output sound anymore
- update sdmmc example to work with new codec
- decrease BLOCK_SIZE to get rid of background noise
- fixe codec driver issue and increase audio buffer

Link: http://cdn.sparkfun.com/datasheets/Dev/Arduino/Shields/WolfsonWM8731.pdf
Link: https://github.com/backtail/libdaisy-rust/commits/5de7f09e82a

Signed-off-by: Olivier Matz <zer0@droids-corp.org>