system: fix pll config after update to stm32h7xx-hal 0.14.0
After hal upgrade in commit b9ff52c00389 ("chore(hal): bump stm32h7xx
hal to 0.14.0 (#34)"), the PLL configuration was triggering a panic:

> panicked at 'Cannot achieve output frequency this low: Maximum PLL divider is 128'

It seems that the HAL is not able to find a valid divider value for
PLL3_Q or PLL3_R. It was probably failing before 0.14.0, but was
silently ignored. Since PLL3_Q and PLL3_R are not used, simply remove
them.

Link: https://www.st.com/content/ccc/resource/training/technical/product_training/group0/81/96/f3/1a/55/d9/40/52/STM32H7-System-Reset_and_clock_control_RCC/files/STM32H7-System-Reset_and_clock_control_RCC.pdf/_jcr_content/translations/en.STM32H7-System-Reset_and_clock_control_RCC.pdf

Signed-off-by: Olivier Matz <zer0@droids-corp.org>